refactor: consolidate imports and remove obsolete orchestrator.py file; enhance context utilities with CurrentDateProvider
