+++ orchestration_engine/__init__.py
@@ -1,16 +1,19 @@
 """Orchestration Engine - Reusable execution engine for SRE automation."""
 
-from .utils.orchestrator_core import OrchestratorCore
+from .utils.orchestrator_core import OrchestratorCore, create_orchestrator_agent, OrchestratorAgentConfig
 from .utils.interfaces import ExecutionContext, PlanningCapableOrchestrator
-from .utils.context_utils import ContextAccumulator
+from .utils.context_utils import ContextAccumulator, CurrentDateProvider
 from .utils.config_manager import ConfigManager
 from .utils.tool_manager import ToolManager
 
 __all__ = [
     'OrchestratorCore',
+    'create_orchestrator_agent',
+    'OrchestratorAgentConfig',
     'ExecutionContext',
     'PlanningCapableOrchestrator',
     'ContextAccumulator',
+    'CurrentDateProvider',
     'ConfigManager',
     'ToolManager'
 ]
